Dodanie zapamietywania poprzedniego katalogu w ktorym byl plik + obsluga bledow na wypadek usuniecie pliku np. gdy pendrive zostanie odlaczony

+++ python_files/functions.py
@@ -34,15 +34,45 @@ def exit(root):
 # Cofa o jedna strone do tylu
 def back(page): 
     show_frame(page)
+    set_default()
+
+def check_path():
+    if var.INITIAL_DIRECTORY.get() =="" or not os.path.isdir(var.INITIAL_DIRECTORY.get()):
+        var.INITIAL_DIRECTORY.set(os.getcwd())
+    else: pass
 
 # Znajduje sciezke do pliku i w zaleznosci od opcji zapisuje w zmiennej globalnej
 def find_file(option):
-    path = filedialog.askopenfilename(title="Find file to encrypt",
-        initialdir=os.getcwd() #otworzy sie w katalagu w ktorym jestesmy
-        )
-    if option==1: var.origin_file_path.set(path)
-    elif option==2: var.encryption_key_file_path.set(path)
+    check_path()
 
+    if option==1: #Dla szukania pliku do szyfrowania/odszyfrowania
+        file_path = filedialog.askopenfilename(title="Find file to encrypt",
+        initialdir=var.INITIAL_DIRECTORY.get()
+        )
+
+        var.origin_file_path.set(file_path)
+
+        # Ustawienie katalogu w ktorym znajdowal sie poprzedni plik jako ten domyslny
+        var.INITIAL_DIRECTORY.set(os.path.dirname(file_path))
+
+    elif option==2: #Dla szukania pliku z kluczem
+        while True:
+            check_path()
+            file_path = filedialog.askopenfilename(title="Find encryption key file",
+            initialdir=var.INITIAL_DIRECTORY.get()
+            ) 
+
+            file_extension = os.path.splitext(file_path)[1] # Wyciaga rozszerzenie pliku
+
+            if file_extension == ".key":
+                var.encryption_key_file_path.set(file_path)
+                var.INITIAL_DIRECTORY.set(os.path.dirname(file_path))
+                break
+            elif file_path=="": # jeżeli użytkownik kliknie "Anuluj" to nie wyskoczy błąd
+                break
+            else:
+                messagebox.showerror("Error", "File with encryption key must be \".key\" extension file")
+            
 def generate_key():
     var.ENCRYPTION_KEY.set(Fernet.generate_key().decode('utf-8'))
     #trzeba zdekodowac klucz bo inaczej bedzie w bajtach
@@ -100,36 +130,40 @@ def encrypt():
     encryption_key_file_path = var.encryption_key_file_path.get().strip()
 
     if origin_file_path != "":
-        if encryption_key_file_path != "":
+        if os.path.isfile(origin_file_path):
+            if encryption_key_file_path != "":
+                if os.path.isfile(encryption_key_file_path):
                 
-                key = "" 
-                plain_text = ""
-                
-                with open(origin_file_path, 'rb') as origin:
-                    plain_text = origin.read()
+                    key = "" 
+                    plain_text = ""
+                    
+                    with open(origin_file_path, 'rb') as origin:
+                        plain_text = origin.read()
 
-                with open(encryption_key_file_path, "rb") as keyfile:
-                    key=keyfile.read()
+                    with open(encryption_key_file_path, "rb") as keyfile:
+                        key=keyfile.read()
 
-                # Tworzy instancję obiektu Fernet, która bedzie uzywana do szyfrowania i deszyfrowania danych
-                f = Fernet(key)
+                    # Tworzy instancję obiektu Fernet, która bedzie uzywana do szyfrowania i deszyfrowania danych
+                    f = Fernet(key)
 
-                file_path = filedialog.asksaveasfilename(
-                    defaultextension=find_extension(origin_file_path),
-                    title = "Zapisz zaszyfrowany plik"
-                )
+                    file_path = filedialog.asksaveasfilename(
+                        defaultextension=find_extension(origin_file_path),
+                        title = "Zapisz zaszyfrowany plik"
+                    )
 
-                if file_path:
-                    with open(file_path, "wb") as enc_file:
-                        encrypted = f.encrypt(plain_text)
-                        enc_file.write(encrypted)
-                        messagebox.showinfo("Good job", "File has been sucessfully encrypted")
+                    if file_path:
+                        with open(file_path, "wb") as enc_file:
+                            encrypted = f.encrypt(plain_text)
+                            enc_file.write(encrypted)
+                            messagebox.showinfo("Good job", "File has been sucessfully encrypted")
 
-                        set_default()
-                else:
-                    messagebox.showerror("Error", "Saving cancelled")
+                            set_default()
+                    else:
+                        messagebox.showerror("Error", "Saving cancelled")
 
-        else: messagebox.showerror("Error", "Choose file with encryption key")
+                else: messagebox.showerror("Error", "Check if \"Encryption key file\" still available on the computer")
+            else: messagebox.showerror("Error", "Choose file with encryption key")
+        else: messagebox.showerror("Error", "Check if \"File to encrypt\" still available on the computer")
     else: messagebox.showerror("Error", "Choose file to encrypt")
